Adding measureTextWidth(), measureTextHeight() & drawText()
The measuring functions calculate the maximum width and height of a given string rendered in the current font.

drawText draws a given string using the current cursor/colors/fonts, automatically wrapped at newlines, but may continue off the screen

drawText also takes an optional string of wrapping characters that be used to determine word breaks to soft wrapping within the current clip rectangle.

An example is included of the string " -" which will soft break at hyphens and spaces.