readRect was failing to return data
My test case for readRect was returning all 0s, so found while rereading
ILI9341 pdf file it looks like we should be passing 0x3f characters in
instead of zeros when we are trying to retrieve the data.

Once I did that things started to work.

Also changed the c variable from uint16_t to unit32_t as it could
potentially hold a value of 320x240x3 which is > 65536 bytes.

Also as per suggesion on forum post, added new define SPICLOCK_READ
2000000
which is used in the SPI.beginTransaction calls when we are doing the
readpixel and readRect calls.