move the #define of swap from header to cpp file
There have been reports of programs not compiling as they may include some other header file that defines swap in conflict with the one here.

And as this one is not used in the header but only in the .cpp file and none of the examples use, it.  It feels safe to simply move it to .cpp file.

graphictest.ino builds fine with this change