New function 'getFont' returns pointer type 'const ILI9341_t3_font_t' of the currently active font in use by the driver.
This can be useful inside functions that make changes to the driver font and then pass back to other code that is not aware that the font has been changed.
Upon entry into such a function we can request the active font and then change back on exit.
eg.

void foobar() {
	const ILI9341_t3_font_t *current_font = tft.getFont();

	// Code

	tft.setFont(*current_font);
}